docker: default docker image name is same as ORFS
the user can specify a locked down version themself for
production.

For local testing, the same docker tag as ORFS is
convenient as that makes compiling a docker image
from ORFS the regular way an obvious alternative
to using "bazel run orfs_env"

This begs the question as to why "bazel run orfs_env"
now exists... Isn't that a CI concern for regression
testing of bazel-orfs and not a user-concern?

Users, such as megaboom, are responsible for publishing
and specifying the docker tag and url or furnishing
the docker image in some out of bounds way(such as
compiling locally).

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>