bazel-orfs: centralize mapping of variables onto stages
The user still can, but no longer has to, know which
argument goes into which stage.

However, the list of this mapping is hardcoded into bazel-orfs.

If an argument is missing from STAGE_ARGS_USES, an error message
is printed that includes:

if stage in STAGE_ARGS_USES[arg]}
Error: key "SDC_FILE" not found in dictionary

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>