optuna: prune to simplest interesting case
multiobjective driven by synthesis parameters is the minimum interesting
case, prune everything else and reduce size of python code and
dependencies in terms of what a reader has to understand.

Fancy dependencies to visualize, if they are easy to use, such as generating
interactive 3d interactive HTML for pareto fron is fine.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>